requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
backtesting>=0.6.0
pandas_ta>=0.3.0
matplotlib>=3.7.0
requests>=2.28.0
PyYAML>=6.0
scipy>=1.10.0
